Moved Runner into its own file
